requirements.txt
asn1crypto==1.5.1
boto3==1.42.57
botocore==1.42.57
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
filelock==3.24.3
fastapi==0.115.0
idna==3.11
jmespath==1.1.0
numpy==2.4.2
packaging==26.0
pandas==3.0.0
platformdirs==4.9.2
pycparser==3.0
pydantic==2.10.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
tomlkit==0.14.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.32.0
